refactor(dev-actions): rename current player component to be dev actions
