+++ client/src/BoardGame.tsx
@@ -4,14 +4,14 @@ import { Local } from 'boardgame.io/multiplayer';
 import { OpenStarTerVillageType } from 'packages/game/src/types';
 import Table from './features/Table/Table';
 import Players from './Players/Players';
-import CurrentPlayer from './CurrentPlayer/CurrentPlayer';
+import DevActions from './DevActions/DevActions';
 
 const Board: React.FC<BoardProps<OpenStarTerVillageType.State.Root>> = (props) => {
   return (
     <div className='Board'>
       <Table {...props} />
       <Players {...props} />
-      <CurrentPlayer {...props} />
+      <DevActions {...props} />
     </div>
   );
 }
